refactor(app)!: refactor project structure refactor folder
BREAKING CHANGE: refactor folder

Closes: refactor project structure
