+++ src/components/material/MaterialExchange.java
@@ -1,6 +1,6 @@
-package components.Material;
+package components.material;
 
-import components.Button.CustomButton;
+import components.button.CustomButton;
 import config.Config;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
