+++ src/config/Config.java
@@ -23,7 +23,7 @@ public class Config {
 	public static String onMouseExitedButtonStyle = "-fx-background-radius: 12px; " + "-fx-background-color: #D9D9D9;";
 
 	public static final int HOMEPANE_HEIGHT = 678;
-	
+
 	public static Background bg(Paint white, CornerRadii radius) {
 		return new Background(new BackgroundFill(white, radius, null));
 	}
