--- src/pane/Untitled.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-
-<AnchorPane fx:id="homePane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Button fx:id="playButton" layoutX="197.0" layoutY="214.0" mnemonicParsing="false" prefHeight="34.0" prefWidth="206.0" text="Play" textAlignment="CENTER" />
-      <Text fx:id="title" layoutX="94.0" layoutY="174.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Dominion's of Realm">
-         <font>
-            <Font size="48.0" />
-         </font>
-      </Text>
-   </children>
-</AnchorPane>
