+++ src/pane/popup/CardPopup.java
@@ -2,8 +2,8 @@ package pane.popup;
 
 import javafx.stage.Popup;
 import pane.ControlPane;
-import components.Button.ExitButton;
-import components.Button.FooterButton;
+import components.button.ExitButton;
+import components.button.FooterButton;
 import config.Config;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
