+++ src/type/BuildingType.java
@@ -1,11 +1,5 @@
 package type;
 
 public enum BuildingType {
-	EMPTYROAD,
- ROAD,
- SUPERROAD,
- HOUSE,
- EMPTYHOUSE,
- TOWER,
- CITY
+	EMPTYROAD, ROAD, SUPERROAD, HOUSE, EMPTYHOUSE, TOWER, CITY
 }
