+++ src/type/CardType.java
@@ -1,13 +1,5 @@
 package type;
 
 public enum CardType {
-	KNIGHT,
-	THIEF,
-	STRONGER,
-	UPGRADE,
-	ROBBER,
-	MOVE,
-	BOMB,
-	ROLL,
-	NUCLEAR,
+	KNIGHT, THIEF, STRONGER, UPGRADE, ROBBER, MOVE, BOMB, ROLL, NUCLEAR,
 }
