+++ src/type/EntityType.java
@@ -1,7 +1,5 @@
 package type;
 
 public enum EntityType {
-	KNIGHT,
-	THIEF,
-	PLAYER,
+	KNIGHT, THIEF, PLAYER,
 }
